+++ src/main/java/com/pos/app/pojo/OrderPojo.java
@@ -13,28 +13,28 @@ import jakarta.validation.constraints.*;
 public class OrderPojo extends AbstractPojo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Integer id;
     
     @NotNull(message = "Product barcode cannot be null")
     @Column(name = "product_barcode")
-    private int productBarcode;
+    private Integer productBarcode;
     
     @NotNull(message = "Client ID cannot be null")
     @Column(name = "client_id")
-    private int clientId;
+    private Integer clientId;
     
     @NotNull(message = "Quantity cannot be null")
     @Min(value = 1, message = "Quantity must be at least 1")
-    private int quantity;
+    private Integer quantity;
     
     @NotNull(message = "Total price cannot be null")
     @Min(value = 0, message = "Total price cannot be negative")
     @Column(name = "total_price")
-    private int totalPrice;
+    private Double totalPrice;
     
     @NotNull(message = "Order ID cannot be null")
     @Column(name = "order_id")
-    private int orderId;
+    private Integer orderId;
 }
